create hash table

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -10,7 +10,7 @@
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *table = NULL;
-	unsigned long int i;
+	unsigned int i;
 
 	table = malloc(sizeof(hash_table_t));
 	if (!table)
